backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
groq>=0.5.0
gunicorn>=21.2.0
